Upstream gradle, libraries and code (#112)
* exodus-android-app: Bump to latest stable gradle version

* exodus-android-app: Bump dependencies to latest stable release

* build.gradle: Bump compileSdk to 31

Drop non-required buildTools version while at it

* build.gradle: Enable Java 11 language features

* build.gradle: Switch to new plugins API

* exodus-android-app: Sync repositories with upstream template

- Remove jcenter as it has been deprecated
- Drop non-required maven
- Add mavencentral support
- Move repositories for all projects into settings.gradle as upstream

* exodusprivacy: Address versionCode deprecation

* NetworkManager: Use NetworkCapabilities for checking internet connection above API 23

* TrackerFragment: Switch to HtmlCompat for HTML parsing

+++ app/src/main/java/org/eu/exodus_privacy/exodusprivacy/adapters/ApplicationViewModel.java
@@ -15,7 +15,7 @@ import java.util.Set;
 public class ApplicationViewModel {
     public String packageName;
     public String versionName;
-    public int versionCode;
+    public Long versionCode;
     public String[] requestedPermissions;
     public @Nullable
     Report report;
